% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateData.R
\name{GenerateData}
\alias{GenerateData}
\title{Title Generate data}
\usage{
GenerateData(
  fEtype.vec,
  wltype.vec,
  Nt = 1000,
  kx.per500 = 100,
  interpolate = TRUE,
  theta = 10,
  maxL = 14,
  exposure.list = NULL,
  std = FALSE,
  verbose = FALSE,
  c1 = NULL,
  c2 = NULL,
  wl.set = NULL,
  fE.set = NULL,
  other
)
}
\arguments{
\item{kw}{}
}
\description{
Title Generate data
}
