% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCV.R
\name{NCV.aceDLNMadditive_fit}
\alias{NCV.aceDLNMadditive_fit}
\title{Title Obtain NCV from the fitted model by aceDLNMadditive}
\usage{
\method{NCV}{aceDLNMadditive_fit}(object, kNCV = 0, NCV.nthreads = 1, verbose = FALSE, ...)
}
\arguments{
\item{object}{object of class \code{aceDLNMadditive_fit}.}

\item{kNCV}{number of neighbors on each side, default \code{0} which is LOOCV.}

\item{NCV.nthreads}{number of threads for NCV calculation, default \code{1}.}

\item{verbose}{whether to print messages during the process, default \code{FALSE}.}

\item{...}{}
}
\description{
Title Obtain NCV from the fitted model by aceDLNMadditive
}
