% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aceDLNMadditive.R
\name{aceDLNMadditive}
\alias{aceDLNMadditive}
\title{Title Fit a Adaptive Cumulative Exposure Distributed Lag Non-Linear Model
(ACE-DLNM)}
\usage{
aceDLNMadditive(
  formula,
  smooth = NULL,
  unpen.smooth = NULL,
  fe.varying = NULL,
  fe.cont = NULL,
  offset = NULL,
  dat = NULL,
  maxL = 14,
  kw = 20,
  kE = 20,
  interpolate = TRUE,
  kx.per500 = 300,
  E.max.list,
  E.min.list,
  ifAIC = FALSE,
  marginalAIC = FALSE,
  ifNCV = FALSE,
  kNCV = 0,
  NCV.nthreads = 1,
  conL = FALSE,
  conLorder = 2,
  pc = NULL,
  par.start,
  par.fix,
  upper.bound,
  CI = 0.95,
  CI.R = 1000,
  CI.seed = 123,
  delta.method = FALSE,
  eta = FALSE,
  GD = TRUE,
  GD.grtol = 1,
  check.BFGS = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{formula}{the formula for the effect of exposure at time t regarding the
exposure history. Use sX() for the distributed lag term. For example y ~
sX(t, x).}

\item{smooth}{the formula for the other smooth functions. The smooths are
specified by mgcv::s. For example ~ s(z). Note also that the Gaussian
random effect models are accommodated, by setting s(bs = "re").}

\item{unpen.smooth}{the formula for the other unpenalized smooth functions.}

\item{fe.varying}{the formula for time-varying covariates in linear
components. For example ~ z}

\item{fe.cont}{the formula for time-invariant covariates in linear
components. For example ~ z}

\item{dat}{the data frame containing the variables.}

\item{maxL}{the maximum lag, default 14.}

\item{kw}{the dimension for B-spline for weight function, default 20.}

\item{kE}{the dimension for B-spline for ACE response-exposure-function,
default 20.}

\item{interpolate}{whether or not to interpolate the exposures (using cubic
B-spline), default \code{TRUE}.}

\item{kx.per500}{the number of knots to fit exposures using B-spline. The
argument works only if \code{interpolate = FALSE}.}

\item{pc}{the point constraint for ACE response-exposure-function, i.e. f(x0)
= 0. Use sum-to-zero constraint if \code{pc = NULL}. Default \code{pc =
NULL}.}

\item{par.start}{the starting values for BFGS in the outer optimization. If
the argument is missing, the starting values are the fitted values from
\code{mgcv::bam}.}

\item{par.fix}{the values of smoothing/dispersion parameters which is fixed
in the outer optimization. For example, set \code{par.fix = c(10,NA,NA)}
for fixing log(theta)=10 if the prior knowledge believes there is no
over-dispersion.}

\item{upper.bound}{the upper bound of parameters in BFGS.}

\item{CI}{the confidence level, default 0.95.}

\item{CI.R}{the number of sampling in the sampling method for CI, default
1000.}

\item{CI.seed}{the random seed in the sampling method for CI, default 123.}

\item{delta.method}{whether or not to use the delta method for CI,default
\code{FALSE}.}

\item{eta}{whether or not to report the CI for the linear predictor eta =
log(mu), default \code{FALSE}.}

\item{GD}{whether or not to use the gradient descent if BFGS fails (max gr >
GD.grtol), default \code{TRUE}.}

\item{GD.grtol}{the tolerance of maximum gradient, default 1. The maximum
numbers of steps for gradient descent is 2. But the BFGS converges in most
cases.}

\item{verbose}{whether or not to print the progress and diagnostic
information, such as the gradients and function values in inner and middle
optimizations.}

\item{E.min}{the lower bound of ACE. The lower bound is set as the minimum of
the exposure if E.min is missing, since this the range we are typically
interested in.}
}
\value{
Object of class \code{aceDLNM_fit}. S3 function \code{summary},
  \code{residuals}, and \code{predict}.
}
\description{
Title Fit a Adaptive Cumulative Exposure Distributed Lag Non-Linear Model
(ACE-DLNM)
}
