% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.aceDLNMadditive_fit}
\alias{predict.aceDLNMadditive_fit}
\title{Title Obtain predictions from the fitted model by aceDLNM}
\usage{
\method{predict}{aceDLNMadditive_fit}(object, dat, CI.fit = TRUE, ...)
}
\arguments{
\item{object}{object of class \code{aceDLNM_fit}.}

\item{dat}{the data frame containing the variables.}

\item{CI.fit}{whether or not to report the CI, default \code{TRUE}.}

\item{...}{}
}
\description{
Title Obtain predictions from the fitted model by aceDLNM
}
